refactor(desktop): extract the omi dot-ring mark for shared chrome
